base code for autosharding
